Formats checking + imports optimization

+++ main.py
@@ -1,14 +1,14 @@
 import os
 import re
+import tempfile
 import zipfile
 from datetime import datetime
-import pyexiv2
-import tempfile
 from tkinter.filedialog import askopenfilename
+
+from pyexiv2 import Image
 
 files_without_date = 0
 
-# Пути хранения файлов
 archive_path = askopenfilename()
 new_archive_path = f"output\\{os.path.basename(archive_path)}"
 
@@ -35,40 +35,43 @@ with zipfile.ZipFile(archive_path, 'r') as archive:
     with zipfile.ZipFile(new_archive_path, 'w') as new_archive:
         for file_info in archive.infolist():
             file_name = file_info.filename
-            with archive.open(file_info) as file:
-                file_data = file.read()
+            if file_name.endswith(('.jpg', '.jpeg', '.png')):
+                with archive.open(file_info) as file:
+                    file_data = file.read()
 
-                date_from_filename = extract_date_from_filename(file_name)
+                    date_from_filename = extract_date_from_filename(file_name)
 
-                if date_from_filename:
-                    # Создаем временный файл для работы с pyexiv2
-                    with tempfile.NamedTemporaryFile(delete=False) as temp_file:
-                        temp_file.write(file_data)
-                        temp_file_path = temp_file.name
+                    if date_from_filename:
+                        # Создаем временный файл для работы с pyexiv2
+                        with tempfile.NamedTemporaryFile(delete=False) as temp_file:
+                            temp_file.write(file_data)
+                            temp_file_path = temp_file.name
 
-                    try:
-                        with pyexiv2.Image(temp_file_path) as img:
-                            metadata = img.read_exif()
+                        try:
+                            with Image(temp_file_path) as img:
+                                metadata = img.read_exif()
 
-                            if 'Exif.Photo.DateTimeOriginal' not in metadata:
-                                formatted_date = date_from_filename.strftime('%Y:%m:%d %H:%M:%S')
-                                metadata.update({
-                                    'Exif.Photo.DateTimeOriginal': formatted_date,
-                                    'Exif.Image.DateTime': formatted_date,
-                                    'Exif.Photo.DateTimeDigitized': formatted_date
-                                })
-                                files_without_date += 1
-                            img.modify_exif(metadata)  # Сохраняем обновленные метаданные
+                                if 'Exif.Photo.DateTimeOriginal' not in metadata:
+                                    formatted_date = date_from_filename.strftime('%Y:%m:%d %H:%M:%S')
+                                    metadata.update({
+                                        'Exif.Photo.DateTimeOriginal': formatted_date,
+                                        'Exif.Image.DateTime': formatted_date,
+                                        'Exif.Photo.DateTimeDigitized': formatted_date
+                                    })
+                                    files_without_date += 1
+                                img.modify_exif(metadata)  # Сохраняем обновленные метаданные
 
-                        with open(temp_file_path, 'rb') as updated_file:
-                            updated_file_data = updated_file.read()
+                            with open(temp_file_path, 'rb') as updated_file:
+                                updated_file_data = updated_file.read()
 
-                        new_archive.writestr(file_info, updated_file_data)  # Добавляем обновленный файл в новый архив
+                            new_archive.writestr(file_info, updated_file_data)  # Добавляем обновленный файл в новый архив
 
-                    finally:
-                        os.remove(temp_file_path)
-                else:
-                    new_archive.writestr(file_info, file_data)  # Добавляем файл в архив без изменений
+                        finally:
+                            os.remove(temp_file_path)
+                    else:
+                        new_archive.writestr(file_info, file_data)  # Добавляем файл в архив без изменений
+            else:
+                new_archive.writestr(file_info, archive.read(file_info))  # Копируем остальные файлы без изменений
 
 print(f"Файлы заархивированы в {os.getcwd()}\\{new_archive_path}")
 print(f"Изначальное количество файлов без даты съемки: {files_without_date}")
